Recoded Kill Aura and switched to a different LWJGL3 implementation.

+++ src/main/java/wtf/bhopper/nonsense/event/EventPriorities.java
@@ -2,10 +2,10 @@ package wtf.bhopper.nonsense.event;
 
 public class EventPriorities {
 
-    public static final byte VERY_LOW = 0;
-    public static final byte LOW = 1;
-    public static final byte MEDIUM = 2;
-    public static final byte HIGH = 3;
-    public static final byte VERY_HIGH = 4;
+    public static final int VERY_LOW = 0;
+    public static final int LOW = 1;
+    public static final int MEDIUM = 2;
+    public static final int HIGH = 3;
+    public static final int VERY_HIGH = 4;
 
 }
